feat(TvCard): refactor to support horizontal layout with updated demo components and styles
