+++ src/assets/scss/style.scss
@@ -128,7 +128,7 @@
   }
 }
 
-.tv-card-vertical {
+.tv-card-horizontal {
   max-width: 600px;
 
   .tv-card {
