+++ src/components/TvCard.vue
@@ -8,7 +8,7 @@ const props = defineProps({
     type: Object,
     default: () => ({}),
   },
-  isVertical: {
+  isHorizontal: {
     type: Boolean,
     default: false,
   },
@@ -20,7 +20,7 @@ const { handleClickLabel, handleClick, handleSecondaryClick, card } = useCard(pr
 </script>
 
 <template>
-  <div class="tv-card-body" :class="{ 'tv-card-vertical': isVertical }" v-if="card">
+  <div class="tv-card-body" :class="{ 'tv-card-horizontal': isHorizontal }" v-if="card">
     <div class="tv-card" :style="card.customStyleCard">
       <div class="tv-card-image">
         <img :src="card.image" :alt="card.alt" />
