+++ src/demo/utils/demos/withHorizontalAndLabels.vue
@@ -2,8 +2,9 @@
   <tv-card
     @click-button="handleButton"
     @click-secondary-button="handleSecondaryButton"
+    @click-label="handleLabelClick"
     :configCard="configCard"
-    is-vertical
+    is-horizontal
   />
 </template>
 
@@ -14,13 +15,19 @@ import '@todovue/tv-card/style.css';
 import '@todovue/tv-label/style.css';
 import '@todovue/tv-button/style.css';
 
+const labels = [
+  { id: 1, name: "Vue.js", color: "#4FC08D" },
+  { id: 2, name: "JavaScript", color: "#F0DB4F" },
+  { id: 3, name: "HTML", color: "#E34F26" },
+];
+
 const configCard = ref({
   title: 'The Creative Routine Before Start Coding',
   description: 'A personal routine I use that helps me generate new ideas and start coding more creatively.',
   alt: 'Card Image',
   image: 'https://res.cloudinary.com/dcdfhi8qz/image/upload/v1763701508/jarnmxxvmhbisvpzzfwu.webp',
   primaryButtonText: 'Read more',
-  secondaryButtonText: 'Save',
+  labels: labels,
 });
 
 const handleButton = () => {
@@ -29,6 +36,10 @@ const handleButton = () => {
 
 const handleSecondaryButton = () => {
   console.log('Secondary button clicked');
+};
+
+const handleLabelClick = (label) => {
+  console.log('Label clicked:', label);
 };
 </script>
